test: move serverless function invocation implementation into common js module shared by both in-process and sandboxed invocations
