+++ tests/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './helpers.js'
 export * from './mock-file-system.js'
-export * from './stream-to-buffer.js'
